MB-3653 Set TAP cursor to the beginning of checkpoint iff necessary
When a given TAP cursor is registered in the checkpoint manager, we
don't need to reset the cursor to the beginning of the checkpoint to
start with if the cursor is already located in that checkpoint. By doing
this, we can reduce the number of items to be sent over network.

However, there is a special registered TAP client (e.g., incremental
backup client) that always requires the cursor to start from the
beginning of the checkpoint to which it currently belongs.

Change-Id: I7fe6fb5824a7fa3f27c8510e71bf3559308e0f32
Reviewed-on: http://review.membase.org/5539
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>