MB3360 - support restoring from WAL'ed db files
If SQLite version is older than 3.7, it won't be able to recover a
WAL-mode database files and report an error of "file is encrypted
or is not a database".

Any datbase file can be deliberately changed out of WAL mode so that
any older version of SLQite can once again access the db file.
    PRAGRAM journal_mode=DELETE;

Change-Id: I3fd040e651b1965008a9e5231509ebbc92eb544b
Reviewed-on: http://review.membase.org/5574
Tested-by: Bin Cui <bin.cui@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>