MB-3678 Return the last closed checkpoint Id to the flusher correctly
There was a bug in returning the last closed checkpoint Id when
the flusher gets all the items for persistence from the checkpoint.
For example, if the checkpoint manager currently has only one
checkpoint with the Id 5, the last closed checkpoint Id to be returned
to the flusher should be 4 instead 0.

Change-Id: Ie65d06d7d1eb6bb969335595deba65b28adca351
Reviewed-on: http://review.membase.org/5618
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>