MB-3684 Don't queue items from invalid backfills into checkpoint
Replica vbucket might receive items from the backfill with an invalid
token even after closing its current checkpoint, because some of items
from the backfill with the invalid token are on the wire even after
that backfill thread is closed.

The above scenario happens especially during the rebalance because the
same named TAP connection is reused for a new connection request with
the same name but a new vbucket filter. In this case, those items are
simply ignored.

Change-Id: Id22fad5b530abb8178c3c4e4992a71a6de06bcb9
Reviewed-on: http://review.membase.org/5625
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>