MB-3700 Allow an active vbucket to receive checkpoint messages
Currently, when a given vbucket is in the active state, it will not
accept checkpoint start/end messages. This will cause a significant
migration problem with older releases which did not enforce
master/slave behavior.

This changes supports
1) A way for the "downstream" master to accept checkpoint begin/end
messages along with sets/gets and any other mutation request.
2) An additional ep_engine parameter to enable this behavior.

It's known that this will lead to inconsistent checkpoints for that
given vbucket between the master and the downstream.

Change-Id: I2f12e1d6e995819e979e7ab40a498316cb39cc59
Reviewed-on: http://review.membase.org/5694
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>