MB-3618 Reset vbucket upon receiving OPAQUE_INITIAL_VB_STREAM message
ns_server is no longer deleting and recreating vbuckets on the
destination node, but instead relying on checkpoints.
However, if the destination node requests too old checkpoints that
are not in memory on the source node, the backfill will be required
from the source node.

If the backfill is required, the destination node should reset the
corresponding vbucket from memory and disk, so that it won't keep
any items that have been already deleted in the source node.

For supporting this, the source node will first send
"TAP_OPAQUE_INITIAL_VBUCKET_STREAM" message to the destination node
if the backfill is necessary, so that the destination node can reset
the target vbucket before receiving the backfill stream.

Change-Id: Idf8c9b7b72fb46f16d9ca917f3e3b17a5d32e86b
Reviewed-on: http://review.membase.org/5761
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>