MB-3746 Create a vbucket if not exist when TAP cursor is registered
It turns out that when we take over a given vbucket to the
destination node, the destination node sometimes receives a new
TAP connection request with checkpoint Ids from eVBucketMigrator
even before that vbucket is instantiated on the destination.

This caused abnormal behaviors on the destination node as we don't
allows TAP cursors to be registered into non-existing vbuckets.
This change resolves the above issue by creating a given vbucket
if not exist upon registering new TAP cursors.

Change-Id: Ib9fa78870cf22cb099581ae0c54fef20c7eea1a7
Reviewed-on: http://review.membase.org/5953
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>