Add a new TAP_OPAQUE message to notify the termination of tap stream
This new message is sent by the eVBucketMigrator to notify that the
source node closes the tap replication stream and switches to
TAKEOVER_VBUCKETS phase.

This is just an informative message and doesn't require any action
from the TAP client. Once the ns_server receives an ACK from the
client, it will close the current replication stream and open the
new stream for TAKEOVER_VBUCKETS.

Change-Id: I71eefb826e44058be4547229b7ad0f777ddcbd1b
Reviewed-on: http://review.membase.org/5978
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>