Include ongoing mutations in vbucket transfer. MB-3736
This addresses a few basic problems that led to data loss.  They are a
bit intertwined, so presenting them as a single changeset makes sense.

1. We always ignored new mutations when in transfer mode.
2. We hung up as soon as we saw the last mutation.
3. We activated the remote end before we were complete.

Regarding #1, there was no distinction between a "dump" and a
"transfer" tap stream.  The former is intended to end approximately at
the state where it began.  It explicitly does not include any new
mutations.  The transfer is similar, but need to stop when no new data
will come in.

We ensure no new data will come in by going into a dead state, and
then ending after we've been in a dead state *and* have completed a
first-level queue transfer from the incoming mutations.  At this
point, there can be no new incoming data and there's nowhere for it to
hide.

The downside of this is that we're not quite able to expedite the
queue transfer at the moment, so this can create a block of up to a
second at the end of any transfer.  I suspect this will be not too
difficult to fix, and enough information to do so is included in this
change, but the goal of this patch is correctness above all.

Change-Id: I2a9c66922a94d804aa179a9302dd2eafb891ff82
Reviewed-on: http://review.membase.org/5989
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>