SYNC replication may return before items get replicated
SYNC replication clients are unblocked when an item is retrieved from
a TAP connection queue. This is not ideal, as the time delta between this
event and the instant the replica server receives the item might be very
significant. With this change set, if the TAP client requests a TAP
connection with explicit ACK support, SYNC replication clients will be
unbocked after receiving an ACK from replication servers - all the TAP
connections created by the vbucketmigrator component (and used for replication)
use explicit ACKs.

This is a fix for the issue MB-3763.

Change-Id: I603deb6730ed74c6143e4429d7acdf1f30de3b07
Reviewed-on: http://review.membase.org/6054
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>