Allow the checkpoint remover to create a checkpoint if required.
The checkpoint remover is periodically scheduled to remove all
closed and unreferenced checkpoints. If the current memory usage
is too high (85% of the max memory size by default), we need to
allow the checkpoint remover to close the current open checkpoint
and create a new checkpoint, because items in the open checkpoint
cannot be evicted from memory. In this way, more closed checkpoints
can be considered by the item pager for eviction.

Change-Id: I0dd80a70d8e249da32bcedecabefff20977f8434
Reviewed-on: http://review.membase.org/6080
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>