MB-3802 Wait until all items are acked for VB transfer completion
We observed the following two data inconsistency issues that can
happen during and right after the rebalance:

1) New item loss: When a given vbucket is transferred from the source
node to the destination node (new master) during rebalance, new items
arrived at the source node right before changing its vbucket state to
dead might not be replicated to the destination as part of vbucket
transfer.

2) New mutation loss: After sending the last item to the destination
for vbucket transfer, the source node does not wait until all the
items are acked from the destination, but instead sends
vbucket_state_active to the destination immediately. Then, the
destination sets its vbucket state to active and will be ready for
receiving new mutations from the clients. However, if there are any
nacked items on the source node, these items will be sent to the
destination again, which can overwrite any new mutations received by
the destination.

This change addresses the above two issues by 1) setting the vbucket
state to dead only when the number of items waiting for acks is under
the given threshold, 2) allowing the tap cursor to get a new mutation
from the dead vbucket, and 3) then sending vbucket_state_active to the
destination after all items are acked from the destination.

Change-Id: I547176577fa98ea803ec1d72b01666bd7b03986e
Reviewed-on: http://review.membase.org/6149
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>