Increase TAP backoff sleep time with larger queue size threshold.
1.7 TAP performs replication faster than 1.6.5 because it does not
need to perform disk fetches. From test results, it turns out that
increasing 1) the persistence queue size threshold on the master and
2) backoff sleep time on the slave show the better behavior especially
during and right after rebalance. However, for the take-over vbucket,
if the take-over operation is about to being completed, we set the
backoff sleep time to 0.5, so that we can reduce the time in the
state of dead vbucket on the source and pending vbucket on the
destination.

Change-Id: I0f0682a3294f430ccbe36988d47895a437b670ca
Reviewed-on: http://review.membase.org/6155
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>