Create a new checkpoint more frequently for high memory usage.
This change allows the checkpoint remover to close the current
open checkpoint and open a new one for a given vbucket if any
of the following condition is true:

1) current memory usage > memory high water mark (75% of bucket
memory size) AND number of items in the open checkpoint >=
MIN_CHECKPOINT_ITEMS (500)
2) current memory usage > 90% of bucket memory size AND
number of items in the open checkpoint > 0

The checkpoint remover is periodically scheduled every five
seconds. In this way, more closed checkpoints can be considered
by the item pager for eviction in high memory usage, and
consequently DGM can be supported in a more flexible way.

Change-Id: I0da7b3f5d67e3691090385ab0af76f0ccde30b05
Reviewed-on: http://review.membase.org/6274
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>