Do not schedule disk flush_all task if it is already queued.
Disk flush_all can be scheduled mutiple times within a very
short period on a given node, if a client issues it multiple times
or TAP_FLUSH is received from other nodes in the cluster.

In this case, we only need to reset the memory hashtable, and do
not need to schedule the disk flush_all if it's been already
scheduled.

Change-Id: I26f9feda0fb18f1550197b0bc94f6f462839ac6b
Reviewed-on: http://review.membase.org/6375
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>