Fix to an issue in creating a new checkpoint in high memory usage.
We observed that the current approach to creating a new checkpoint
in high memory usage (e.g., beyond high watermark) caused the
checkpoint manager to instantiate too many new checkpoints, which
can consume even more memory if the persistence/tap cursors are
slow and much more behind the current open checkpoint.

This change fixes the above issue by creating a new checkpoint in
high memory usage, iff all cursors are currently in the
current open checkpoint and the number of items in the current
open checkpoint is more than the specified minimum number.

Note that items belonging to the current open checkpoint cannot be
evicted from memory as they are currently (or will be) visited by
the persistence/tap cursors. Therefore, if the persistence and tap
cursor are fast and currently in the open checkpoint, it would be
better to close the current checkpoint and open the new one in case
of high memory usage, so that those cursors can move to the new
checkpoint quickly and items in the closed checkpoint can be purged
from memory.

Change-Id: Iccadd2d665cbf0577944e9530898aed4d4960c17
Reviewed-on: http://review.membase.org/6448
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>