MB-3887 Create a new checkpoint if necessary during backfill phase
Before this change, we didn't create a new checkpoint on the
destination node while receiving backfill streams from the source.
This means that all the items from backfill will be pushed into the
current open checkpoint with checkpoint Id 1. This can causes OOM on
the destination during the rebalance, as we cannot evict any items
from the current open checkpoint.

To fix the above issue, this change basically allows the destination
to create a new checkpoint based on the number of items on the
current open checkpoint. When the destination receives the
backfill_end message, it won't be allowed to create any new
checkpoints, which means that the creation of a new checkpoint is
totally controlled by the source node.

Change-Id: I5a19f46313fd7e0c5c369e4204e912f69140b31e
Reviewed-on: http://review.membase.org/6507
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>