Don't create a checkpoint if the checkpoint with the same ID exists
When the replica vbucket receives CHECKPOINT_START message, this
change does not allow the replica vbucket to create a new
checkpoint if the checkpoint with the same ID already exists. But,
it still removes all the checkpoints with their IDs greater than
the ID of the checkpoint to be created.

Change-Id: I8d09e82fdfcc9c066ed2e8360ae60128cfb88d69
Reviewed-on: http://review.membase.org/6519
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>