Put memory threshold on the bucket quota in accepting a new mutation
Checkpoint-based persistence and replication can consume more
memory if persistence or TAP cursors become slow. Consequently, we
observed that if we load a large number of items in a short time,
item evictions are performed well until 2X DGM, but not anymore with
more than 3X DGM.

We realized that ep_engine previously accepted any mutations as long
as its current memory usage is less than the bucket memory quota.
This can easily cause permanent OOM if persistence or replication is
getting slow. To resolve this issue, we put memory threshold on the
bucket quota in accepting a new mutation. The default threshold is
90% of the bucket memory quota, but is configurable by passing it
as an engine parameter. We also assume that the client will do
temporary backoff if it receives temp OOM errors from the server.

Change-Id: Ie5c1f503a6d315efe7cf790b3a4f8a177f73729c
Reviewed-on: http://review.couchbase.org/6643
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>