Adjust the last persisted checkpoint ID if the cursor is repositioned
When the replica vbucket receives CHECKPOINT_START message and
purges all the checkpoints with IDs greater than the new checkpoint
ID, it might reposition the persistence cursor to the beginning of
the new checkpoint if the cursor was walking through those removed
checkpoints. In this case, we should also set the ID of the last
persisted checkpoint to (new checkpoint ID -1) in the vbucket table.

Change-Id: I217cf7140b040aed1b879801c0cd335655d07139
Reviewed-on: http://review.couchbase.org/6653
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>