keep cookie refcount for backfill thread life duration.
When new tap connection replaces old with same name we release cookie,
which will likely free it and allow memcached to reuse same cookie for
later connections. Which will make validity check in backfill thread
to succeed when backfill thread is in fact stale. This was observed in
'real-world' by Chiyoung Seo.

In order to prevent that we just increment refcount before starting
backfill thread and decrement it when backfill thread is dying.

Because memcached refcount ops implementation is not synchronized and
refcount decrementing from tap thread and backfill thread may race
each other, we have to use some own lock. TapConnMap#notifyLock was
choosen.

Change-Id: If640823152e1d0912f5f038b2a0e0ac4cd3e04b3
Reviewed-on: http://review.couchbase.org/6665
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>