MB-3935 Destroy SyncListeners from the NIO Dispatcher
As seen in MB-3935, we can have an NIO dispatcher thread making use of
a SyncListener while a worker thread is deleting it.

In this case, a lock was acquired by the dispatcher thread, some work
was begun on the SyncListener, and then the engine decided it was done
with the SyncListener, so it deleted it.  At the end of the
destructor, the Mutex destructor was called which called
pthread_mutex_delete.  Meanwhile, the NIO dispatcher was still
*holding* that lock, causing pthread_mutex_delete to fail, a condition
on which we abort.

This change pushes a job into the NIO dispatcher to have it perform
the deletion, preventing the deletion from running concurrently with
any other action on it.

Change-Id: I602d6eb4c739b5bf396d5a631163bab0a88b5ea8
Reviewed-on: http://review.couchbase.org/6741
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>
Reviewed-by: Farshid Ghods <farshid.ghods@gmail.com>
Tested-by: Farshid Ghods <farshid.ghods@gmail.com>