Allow duplicate INITIAL_VB_STREAM messages to the destination.
As ns_server makes a new connection repeatedly with the same name and
new list of vbuckets during the rebalance, INITIAL_VB_STREAM message
for a given vbucket might not be sent to the destination if we don't
allow duplciates on that message. The reason was that if a new
connection request was made within a very short time, the old
connection coudln't even send INITIAL_VB_STREAM message to the
destination.

Change-Id: Icf12cbf6508115badd4c0e23624aa76e490bb930
Reviewed-on: http://review.couchbase.org/6756
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>