Check backfill thread validity by TAP connection's name.
From 1.7, replication is still going during rebalance, and
therefore, the source node will receive a new connection
request for the same named TAP with the list of existing
+ a new vbucket every time the new vbucket's take-over operation
is completed during rebalance.

Whenever the source node processes a new connection request,
it extracts the list of vbuckets that require backfill.
This results in creating bunch of backfill threads
with many duplicate vbuckets. We observed that this behavior
caused lots of disk IO, slow or stuck rebalance, and data
loss due to race conditions.

To resolve this issue, we provided the fix
(http://review.couchbase.org/#change,6845), which schedules a
backfill thread for new vbuckets only for a given named TAP
connection.  This change allows us to check backfill thread
validity by TAP connetion name and its connectivity status.

Change-Id: I0b5ed5574727eb724c7247f44cfb1e498d163888
Reviewed-on: http://review.couchbase.org/6929
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>