Consider memory usage in creating a new checkpoint by TAP consumer
If 1.7 TAP consumer connects to TAP producer that does not support
checkpoint operations, it should create a new checkpoint for itself
and consider the current memory usage in creating a new checkpoint.

If the current memory usage is too high, TAP consumer triggers the
creation of a new checkpoint, so that the item pager can purge
items from the old open checkpoint.

Change-Id: I10a4daf5033e78d9743ae297870355992bc9fbba
Reviewed-on: http://review.couchbase.org/7015
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>