Send a backfill completion message per vbucket to the destination
As a given TAP consumer instance might be closed and instantiated
again to connect to the named TAP producer, we should not maintain
the list of backfilled vbuckets in the TAP consumer instance.
For this, the named TAP producer should send a backfill completion
message per vbucket to the TAP consumer.

Change-Id: Id31982cf00b49fa1f553ed47915b6060e88b6fec
Reviewed-on: http://review.couchbase.org/7027
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Bin Cui <bin.cui@gmail.com>