Ensure client libevent is shut down when shutting down the engine.
If we teardown and deallocate a connection while the object is still
registered with libevent, an event could fire and pass our invalid
object to our method where we just kind of stomp on memory and do
really dumb stuff.

It's not *entirely* clear that this happened since I can't figure out
how to make it happen myself, but the evidence of this having occurred
is substantial and circumstantial.

Change-Id: I5302b9055d73395efceda1d243252aa792fb40a7
Reviewed-on: http://review.couchbase.org/7137
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>