Clear engine specific data before SyncListener object destruction
Currently, we attempt to clear engine specific data within the
SyncDestruction callback method. This, however, may introduce a race
condition as a result of which the cookie whose engine specific data
is being cleared may get deleted in a separate thread before it is
accessed in the callback method. The race condition was exposed by a
few failing ep-engine unit tests.

This change set avoids the race condition by clearing the engine
specific data before destruction of the SyncListener object is
dispatched.

Change-Id: I3d6b09adf4e8118918053156a0bd90a8b80cf440
Reviewed-on: http://review.couchbase.org/7602
Tested-by: Dustin Sallings <dustin@spy.net>
Reviewed-by: Dustin Sallings <dustin@spy.net>