Fix failing ep-engine atomic_queue_test
Simplified the logic to join the worker threads created by the main
thread. Essentially, since we do not care about the actual return
value from the threads, it is alright to return a NULL from the
threads and throw it away in the main thread. If at all we must
assert some condition, we should do so in the worker thread. The
previous method of calling pthread_join() was causing a segfault
due to improper casting of the return value.

Change-Id: I244915ff556518bfb470f12d6716cf26c1a1c27e
Reviewed-on: http://review.couchbase.org/7730
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Dustin Sallings <dustin@spy.net>