Allow mc-engine to abort clean shutdown
Since ns_server doesn't wait for memcached to exit before terminating
couchdb is terminated before we could persist all of our changes.
This means that memcached would be trying to connect over and over
again to couchdb in order to persist the data, but the process
is no longer there. This commit allows us to terminate that
sequence by setting the variable:

allow_data_loss_during_shutdown=true

as an engine parameter. When this variable is set mc-engine will
check the existence of the parent process between each attempt
to connect to mccouch and exit with exitcode 1 if the parent
is gone.

Change-Id: I552d75de89672810c4358c25606ce0ec1ea3ca9c
Reviewed-on: http://review.couchbase.org/7756
Reviewed-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>
Tested-by: Aliaksey Kandratsenka <alkondratenko@gmail.com>