Ensure mc can process database state transitions before writing data.
SQLite has all vbucket tables ready all of the time.  In practice,
this means that SQLite doesn't care, but it's also part of the reason
that we have to be really careful when reusing them.

In CouchDB, we create and destroy databases, so we want to ensure that
we see these events in the operation stream at the right times.

Change-Id: I3d24625edecdf58113a99df8ac75170e6a75c5f3
Reviewed-on: http://review.couchbase.org/7832
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>
Tested-by: Srinivas Vadlamani <srinivas@couchbase.com>