Set the vbucket states on the backend before setting data
Commit 4082a7bd47bc4692a2fc3707e9210f11e99bd76c added a queued item
to the write queue with the state change of a vbucket, but the write
queue is only to be used by the flusher thread (it is used without
locking). To avoid adding more locks, this patch will "cache" the
state of the vbuckets and update the state of the vbuckets if it
detects a change.

Change-Id: Idb7abddb39fda1d5f62dffd6696c1d0c3658ec31
Reviewed-on: http://review.couchbase.org/7860
Tested-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Filipe David Borba Manana <fdmanana@gmail.com>