Memcached will always send error messages for tap notifications
Initially tap streams wouldn't send anything back, and we added
the nack mode to be backwards compatible. The old mode should not
appear in the upstream memcached (because it's a new interface
there), so it should not be present in our code as well.
The bucket migrators supports getting errors, so this change
should work for our product.

Change-Id: I31b39838b00506acdf021c394d87a9ad77faefdb
Reviewed-on: http://review.couchbase.org/7982
Reviewed-by: Bin Cui <bin.cui@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Trond Norbye <trond.norbye@gmail.com>