Put the limit on the max number of checkpoints per vbucket.
If there are any slow cursors that still walk through the closed
checkpoints, we should NOT close the current open checkpoint and
create a new checkpoint frequently. Otherwise, it will cause more
memory overhead due to less deduplication and more closed
checkpoints.

The default max number of checkpoints per vbucket is currently set
to two (i.e., last closed checkpoint and current open checkpoint).
This means that even if any conditions (e.g., time-bound or number
of items in the open checkpoint) for creating a new checkpoint is
satisfied, a new checkpoint won't be created if there are any
cursors that still walk through the last closed checkpoint.

Note that all the checkpoint-related config parameters are
non-static variables from 2.0 to support different values in
multiple buckets.

Change-Id: I5c0c05607d56ee4520941394b43b3cdbb0668e90
Reviewed-on: http://review.couchbase.org/9358
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>