MB-4202 Avoid using std::string in prepend/append operations.
We observed that as C++ STL and string classes use their own
memory pool allocators, it sometimes shows a quite bad behavior
in its memory pooling. Especially, the current prepend/append
APIs use std::string as temporary storage, which can result in
a number of destructed objects that are not immediately freed
and returned back to OS, if the client invokes prepend/append
APIs frequently with large values.

To mitigate this issue, we temporarily allocate a memory segment
through the underlying memory allocator (e.g., tcmalloc in linux),
and then explictly free it at the end of prepend/append
operations.

Change-Id: I6b128e437b3ffbcc9d981f8aa215324c1d89b7e8
Reviewed-on: http://review.couchbase.org/9359
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>