Pause the expiry pager if the persistence queue is too large
We observed that if there are lots of items that expired at
the same time, it will cause the persistence queue to grow
significantly in a very short time. Consequently, this causes
other write operations (e.g., set, add) to wait in the queue
for a long time. As write operations are more important,
deleting expired items should be backed off if the persistence
queue is too large.

Change-Id: I435edcb9e681edf926aadd3d1a7d2d83a7aec8de
Reviewed-on: http://review.couchbase.org/9361
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Reviewed-by: Srinivas Vadlamani <srinivas@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>