[Backport] Further deduplication in writing items from checkpoints
When the flusher persists an item from a checkpoint, it can grab
the itme's latest value from the memory hashtable if there were
any new mutations for that item, instead of using the value
retrieved from the checkpoint. Then, the flusher can mark the item
as "non-dirty" upon persisting it with the latest value.

Meanwhile, when we push the list of items from a checkpoint into
the flusher's queue, we should check if an item is dirty or not,
and push only dirty items into the flusher. This change results
in reducing disk write overhead by 10 - 20% in our tests.

Change-Id: I6e87b4bc5bc4e91ff29ba8248bd2e08dcdc342fe
Reviewed-on: http://review.couchbase.org/11810
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>