Don't create a copy of the key
From looking at a performance metric we spent more time creating
the object than doing the search for the item with the key on
the next line in the source.

From the test I ran it was responsible for ~1.2 CPU sec, and without
it I'm getting 0.2. It's not much; the total CPU time for the function
was ~30 sec, but there is no point of calling allocating memory and
copy the content just to pass the string on to the search..

Change-Id: I200ddb7e0f26bc6b12e2c155e24471cc1141853e
Reviewed-on: http://review.couchbase.org/9876
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>