Implement the observe registry
This the implementation of the basic observe registry. This
code makes it possible for the observe and unobserve commands
to store keys into Couchbase. This code does not include the
ability for observe commands to be cleaned up or callbacks that
will change the statistics for each key that is being observed.
Those changes will come in future commits.

Change-Id: I25502bed66afcd5eaa7f5da465805abf1a2b8938
Reviewed-on: http://review.couchbase.org/9957
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>