MB-4305 Synchronize accesses to the suspended flag through a lock.
From the large cluster tests with hundreds of millions of items,
we saw the race condition in accessing the suspended flag in a
TAP producer, which resulted in suspending the TAP producer
forever. To resolve this issue, this suspended flag is now
accessed through a global lock per TAP producer.

Change-Id: I8caa096a81cf77832d3be6c036a3fd460b1b9195
Reviewed-on: http://review.couchbase.org/10801
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>