Disabled notifications to TapConnMap by default
During _normal_ load the server will receive data mutations causing the
system to process events all the time. Grabbing the mutex just to send
a notification will only lead to contention on the mutex.

To avoid the thread to busy-loop we're going to do a 1 sec sleep if
there isn't _any_ work to be performed.

Change-Id: Id47e0b6cc5d6678d6db1b16e0fe00811cacfa28d
Reviewed-on: http://review.couchbase.org/10320
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>