When a key is mutated change the deleted field to false
We want to do this because we only want to show that a key
is in the deleted state if it doesn't exist in Couchbase.
Lets say a key is being watched for multiple cas values
and then that key gets deleted. We will record deleted
on all of that key's observed cas values. If a user later
sets a key in the database with the same key name then the
deleted state will be cleared and replaced with mutated.

Change-Id: I53337ba355d7196a6f7234b9990125e761af51eb
Reviewed-on: http://review.couchbase.org/10374
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Michael Wiederhold <mike@couchbase.com>