MB-4402 Do not pass disk backfill dispatcher callback to kvstore.
As a disk backfill dispatcher task is inherited from both
DispatcherCallback and Callback<GetValue> classes, we should not
cast a disk backfill task to shared_ptr<Callback<GetValue>> and
pass the casted instance to the underlying kvstore. This causes
the backfill dispatcher task instance to be freed from memory
before terminated by the dispatcher.

To resolve this issue, a separate backfill callback class is
created for the kvstore. This fix resolves three rebalances
failures (MB-4402, MB-4394, MB-4381) that were seen in the
functional rebalancing tests.

Change-Id: I2c0a17b2b82e0f3222e2e800cd216589a25bdf04
Reviewed-on: http://review.couchbase.org/10652
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>