MB-4430: VBMigrator dies when tap is registered with same name
The issue here actually occurs when when we try to create a tap
connection to when a previous connection with the same name is
already open. What we were doing was moving the first tap
conneciton onto the second tap connection. This meant that when
the second tap connection closed it would also close the first
tap connection.

The fix that was put in for this is to make it so the registered
tap streams will not accept new tap streams if the name specified
is already being used in a stream.

Change-Id: Iae86d19da7df0512be62369b010eb705b9e81fb8
Reviewed-on: http://review.couchbase.org/10916
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>