MB-4417 Ignore duplicate checkpoint messages in incremental backup
If tap_keep_alive is given as t, then the server will keep a TAP
connection for t sec even if the client is disconnected. If the
client reconnects within t sec, the server will reuse that
connection. This named TAP behavior sometimes causes the
incremental backup client to receive duplicate checkpoint
messages. In this case, those duplicate messages should be simply
ignored.

Change-Id: I37394522597d4231da8db97c1762e3be8879f7d4
Reviewed-on: http://review.couchbase.org/10924
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>