MB-4437 Registered TAP can pass a checkpoint Id to the master.
This change allows the named TAP registration script to pass the
last closed checkpoint id from the slave to the master so that
the master can set the named TAP's cursor position to the
beginning of the checkpoint following that closed checkpoint.

For example,
mbadm-tap-registration -h 127.0.0.1:11210 -r slave1 -l 1

The above command registers the named TAP with the name "slave1"
and the last closed checkpoint Id "1".

Change-Id: I945a8fe0318e3bd3e5024b0c72f35c2fa63c7b35
Reviewed-on: http://review.couchbase.org/10969
Reviewed-by: Steve Yen <steve.yen@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>