Notify a new vbucket batch count through the RW dispatcher.
Previously, the memcached frontend thread notifies a new vbucket
batch count to mccouch through the socket. But, this caused the
race condition with the RW dispatcher on the socket buffer because
there is no lock to synchronize accesses to the socket buffer.

To resolve this issue, a new vbucket batch cound is notified to
the mccouch by the RW dispatcher.

Change-Id: I83000e8defd50608bc56c824b61b4a857f90c978
Reviewed-on: http://review.couchbase.org/11119
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>