Fix bug in how StoredValue size is computed
There were two issues in how the StoredValue size was
computed.

First, the alignment that is added to the end
of the key and value will be computed wrong if no
alignment is needed. In this case the alignment would be
set to 8.

Second, adding in the size of the value_t struct is
redundent since it is already accounted for in the
sizeof(StoredValue) call. (Located in the sizeOf
function.

Change-Id: Idd3aba5f141559bed93f378334905c02d0837f24
Reviewed-on: http://review.couchbase.org/11259
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>