MB-4359: Notify all pending connections through a separate thread
This bug was causing get and set operaitont o take a long time
during rebalancing a 25 node cluster. The cause of the bug has
to do with vbucket state changes not happening quickly due to
other jobs on the nio dispatcher taking a long time. The result
is a temporary drop in performance.

This fix moves the vbucket state change job off of the nio
dispatcher and onto the tap notify thread.

Change-Id: Icdfd28ffa2b3c48f3b07f384062d161caf227086
Reviewed-on: http://review.couchbase.org/11307
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Reviewed-by: Michael Wiederhold <mike@couchbase.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>