[Backport] MB-4292: Always add item to dirty queue during delete
This fixes a race condition that causes items to not be expired
on replica nodes. The race is caused when the flusher removes
an expired item from disk and before the expiry pager removes
the item from memory a tap delete message is recieved from the
source node.

The fix causes a deleted item to always be added to the flushQueue
and if the item is already removed from disk the persistence
callback will be immediately called.

Change-Id: If83a9160308a1d22768010901d3879bbdb3885c6
Reviewed-on: http://review.couchbase.org/11657
Reviewed-by: Chiyoung Seo <chiyoung.seo@gmail.com>
Tested-by: Chiyoung Seo <chiyoung.seo@gmail.com>